Remove `DbTest`
